int -> size_t

+++ src/losses.h
@@ -9,26 +9,26 @@
 class CrossEntropyLoss {
   public:
     std::pair<double, Tensor> operator()(const Tensor &logits, const Tensor &labels) const {
-        int batch_size = logits.shape[0];
-        int num_classes = logits.shape[1];
+        size_t batch_size = logits.shape[0];
+        size_t num_classes = logits.shape[1];
         Tensor flattened_labels = labels.squeeze();
 
         double total_loss = 0.0;
         Tensor grad = Tensor::zeros({batch_size, num_classes});
 
         // Compute loss and gradients per sample
-        for (int i = 0; i < batch_size; ++i) {
-            int true_class = static_cast<int>(flattened_labels.at({i}));
+        for (size_t i = 0; i < batch_size; ++i) {
+            size_t true_class = static_cast<size_t>(flattened_labels.at({i}));
 
             // Find max logit for numerical stability
             double max_logit = logits.at({i, 0});
-            for (int c = 1; c < num_classes; ++c) {
+            for (size_t c = 1; c < num_classes; ++c) {
                 max_logit = std::max(max_logit, logits.at({i, c}));
             }
 
             // Compute log_sum_exp for this sample
             double sum_exp = 0.0;
-            for (int c = 0; c < num_classes; ++c) {
+            for (size_t c = 0; c < num_classes; ++c) {
                 sum_exp += std::exp(logits.at({i, c}) - max_logit);
             }
             double log_sum_exp = std::log(sum_exp);
@@ -38,7 +38,7 @@ class CrossEntropyLoss {
             total_loss += sample_loss;
 
             // Gradient: softmax - one_hot
-            for (int c = 0; c < num_classes; ++c) {
+            for (size_t c = 0; c < num_classes; ++c) {
                 double softmax_c = std::exp(logits.at({i, c}) - max_logit - log_sum_exp);
                 double one_hot_c = (c == true_class) ? 1.0 : 0.0;
                 grad.at({i, c}) = (softmax_c - one_hot_c) / batch_size;
